ASoC: msm8x10-wcd: Reduce pop noise when earpiece is switched fast
In some scenarios, earpiece will be switched on just after turned off,
which will cause the audible pop noise in earpiece.
And if customer increases the EAR PA gain, this noise can be much louder.
So enable the SND_SOC_DAPM_POST_PMD event for EAR PA kcontrol
to stablize the PA.

Change-Id: I8991e268e5d161ab14e19ef4bb7e60211758c4f2
CRs-fixed: 591848
Signed-off-by: Walter Yang <yandongy@codeaurora.org>