thermal: qpnp-adc-tm: Support timer selection
Existing channels use a single timer to select the
measurement interval polling rate. Changing the interval
polling value for this timer affects the polling interval
across all the channels.

Provide an option for clients to choose from 3 of the
supported timers. Existing support has each timer linked
to a pre-configured interval polling rate. Clients can
add a optional property if it needs a specific interval
polling rate on timer2 and timer3. If the property is not
present the channel defaults to timer1 and its associated
interval polling rate.

Change-Id: Ia3c7f13081dcddb901ef446cc776f524f875fba2
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>

Conflicts:
	arch/arm/boot/dts/msm-pm8941.dtsi
	arch/arm/boot/dts/msm8226.dtsi
	arch/arm/boot/dts/msm8610.dtsi